Create linearProbing.java
This program simulates the operations of a HashTable that uses Linear Probing. 

The first line (= 5 in the example) is the initial size of the hash table. The second line (= 12 in the example) indicates the number of commands you have to conduct to the hash table. The commands include “insert” (= insert a key to the table), “displayStatus” (= display the status of an entry in the table), “tableSize” (= display the size of the table), “search” (= search a key in the table) and “delete” (= delete a key in the table).

Note that if the load factor becomes greater than 0.5 after a new insert, you have to conduct the rehashing. In other words, you have to find the first prime number that is twice as large as the current table size and move the valid keys in the current table to the new table. After that, you have to insert the new key value. This is the result after the “insert 20” command.

<b>Input format: This is a sample input from a user.</b> 

5
12
insert 17
insert 12
displayStatus 2
tableSize 
insert 20
tableSize
search 20
search 15
displayStatus 1
delete 12
displayStatus 1
displayStatus 2

This is the correct output. For the “displayStatus” command, your program should display the status of an entry of the table. For example, your program should display “17 Active” for the first “displayStatus 2” command. For the second “displayStatus 2” command, it should display “Empty”.

17 Active
5
11
20 Found
15 Not found
12 Active
12 Deleted
Empty

<b>Input format: This is a sample input from a user.</b> 

7
8
insert 100
insert 16
insert 37
delete 16
displayStatus 3
insert 72
displayStatus 2
displayStatus 3

This is the correct output. 

16 Deleted
100 Active
72 Active